Preventing errors when buying items in store
Thought the double tap prevention variable wasn't getting reset properly, so using a timeout to reset it.
Not sure why a store item may not be properly built for buying, but it's probably a timing thing.
Whatever the reason for these things, a little double-checking to make sure to ignore and rectify such errors is necessary.
This fixes Issue #15